Homework #13 + #14

+++ src/main/scala/Event.scala
@@ -1,9 +1,9 @@
 object Event {
   sealed trait Event
 
-  case class Added(id: Int, amount: Int) extends Event
+  case class Added(id: Int, amount: Double) extends Event
 
-  case class Multiplied(id: Int, amount: Int) extends Event
+  case class Multiplied(id: Int, amount: Double) extends Event
 
-  case class Divided(id: Int, amount: Int) extends Event
+  case class Divided(id: Int, amount: Double) extends Event
 }
